requirements.txt
apache-airflow==2.8.4
numpy==2.0.2